+++ .github/workflows/archive.yaml
@@ -22,4 +22,4 @@ jobs:
       - name: Install packages
         run: yarn install
       - name: Run archive
-        run: yarn start
+        run: yarn archive
